+++ lab/reports/markup.py
@@ -65,6 +65,10 @@ CSS = """\
 """ % globals()
 
 
+def escape(text):
+    return "''%s''" % text
+
+
 def _get_config(target):
 
     config = {}